feat(routing): directly link content records
Instead of using a generic InternalLink component for better content editing experience.
