+++ datocms-environment.ts
@@ -3,5 +3,5 @@
  * @see docs/getting-started.md on how to use this file
  * @see docs/decision-log/2023-10-24-datocms-env-file.md on why file is preferred over env vars
  */
-export const datocmsEnvironment = 'page-model-names';
+export const datocmsEnvironment = '160-link-records';
 export const datocmsBuildTriggerId = '30535';
