--- src/blocks/InternalLink/InternalLink.astro
@@ -1,25 +1,0 @@
----
-import type {
-  InternalLinkFragment,
-  SiteLocale,
-} from '@lib/types/datocms';
-import Link from '@components/Link/Link.astro';
-import { getHref } from '@lib/routing';
-
-export interface Props {
-  link: InternalLinkFragment;
-  openInNewTab?: boolean;
-}
-type Params = {
-  locale: SiteLocale;
-};
-
-const { link, openInNewTab, ...props } = Astro.props;
-const { locale } = Astro.params as Params;
-const { title, page: record } = link;
-
-const href = getHref({ locale, record });
-
----
-
-<Link {href} {openInNewTab} {...props}><slot>{title}</slot></Link>
