--- src/blocks/InternalLink/InternalLink.fragment.graphql
@@ -1,14 +1,0 @@
-#import '@lib/routing/PageRoute.fragment.graphql'
-
-fragment InternalLink on InternalLinkRecord {
-  title
-  page {
-    __typename
-    ... on PageRecord {
-      ...PageRoute
-    }
-    ... on HomePageRecord {
-      title
-    }
-  }
-}
