+++ src/blocks/TextBlock/README.md
@@ -9,5 +9,5 @@
 
 - Renders a rich text component as HTML, with support for headings, lists, bold, italic, emphasized text, etc.
 - Renders nested blocks using [`<Blocks />`](../Blocks.astro), so it supports any Block.
-- Renders internal links using [`<InternalLink />`](../InternalLink/InternalLink.astro).
+- Renders internal links using [`<LinkToRecord />`](../../components/LinkToRecord/LinkToRecord.astro).
 - Renders headings with a custom component to avoid having multiple H1's on the page.
