+++ src/blocks/TextBlock/TextBlock.fragment.graphql
@@ -1,8 +1,9 @@
 #import '@blocks/ImageBlock/ImageBlock.fragment.graphql'
-#import '@blocks/InternalLink/InternalLink.fragment.graphql'
 #import '@blocks/TableBlock/TableBlock.fragment.graphql'
 #import '@blocks/VideoBlock/VideoBlock.fragment.graphql'
 #import '@blocks/VideoEmbedBlock/VideoEmbedBlock.fragment.graphql'
+#import '@lib/routing/HomeRoute.fragment.graphql'
+#import '@lib/routing/PageRoute.fragment.graphql'
 
 fragment TextBlock on TextBlockRecord {
   text {
@@ -22,8 +23,13 @@ fragment TextBlock on TextBlockRecord {
       }
     }
     links {
-      id
-      ...InternalLink
+      __typename
+      ... on HomePageRecord {
+        ...HomeRoute
+      }
+      ... on PageRecord {
+        ...PageRoute
+      }
     }
     value
   }
