+++ src/blocks/TextImageBlock/TextImageBlock.fragment.graphql
@@ -1,7 +1,8 @@
 #import '@blocks/ImageBlock/ImageBlock.fragment.graphql'
-#import '@blocks/InternalLink/InternalLink.fragment.graphql'
 #import '@blocks/TableBlock/TableBlock.fragment.graphql'
 #import '@blocks/VideoEmbedBlock/VideoEmbedBlock.fragment.graphql'
+#import '@lib/routing/HomeRoute.fragment.graphql'
+#import '@lib/routing/PageRoute.fragment.graphql'
 
 fragment TextImageBlock on TextImageBlockRecord {
   text {
@@ -18,8 +19,13 @@ fragment TextImageBlock on TextImageBlockRecord {
       }
     }
     links {
-      id
-      ...InternalLink
+      __typename
+      ... on HomePageRecord {
+        ...HomeRoute
+      }
+      ... on PageRecord {
+        ...PageRoute
+      }
     }
     value
   }
