+++ src/components/LinkToRecord/README.md
@@ -1,0 +1,7 @@
+# LinkToRecord
+
+**Resolves URL for different page records from the CMS (like the home page or a generic page).**
+
+## Usage
+
+See [documentation on Internal links to records](../../../docs/routing.md#internal-links-to-records)
