+++ src/lib/routing/README.md
@@ -1,0 +1,3 @@
+# Routing
+
+See [documentation on Routing](../../docs/routing.md).
